Updated python and css styling of image pairs
Python script now will not return an error calling hist
on a stock that has not been incepted at any point in the timeframe.
HTML images were changed to actually reflect the project, CSS styled
the pairs to make it even.

Testing with a test python file to create a RSI and MACD without FLASK